desloppify: resolve render-boundary-consolidation cluster by extracting prepare.py and expansion.py from orchestrator.py
